opus编解码调用测试

+++ app/src/main/java/lee/com/audiotalkie/model/RecordDataCallback.java
@@ -7,6 +7,6 @@ package lee.com.audiotalkie.model;
  */
 public interface RecordDataCallback {
 
-    void data(byte[] data);
+    void data(short[] data);
 
 }
